change training from incr. to reevaluate top l

+++ src/experiments/permus/FitnessFunction_permus.h
@@ -1,3 +1,3 @@
 #include "cpunetwork.h"
 
-float FitnessFunction_permu(NEAT::CpuNetwork *net, int n_evals);
+double FitnessFunction_permu(NEAT::CpuNetwork *net, int n_evals);
