+++ src/experiments/permus/permu_problem_src/LOP.h
@@ -40,7 +40,7 @@ public:
     /*
      * Entries matrix of the LOP.
      */
-	float ** m_matrix;
+	double ** m_matrix;
     
 	/*
 	 * The size of the problem.
@@ -61,14 +61,14 @@ public:
 
 
 
-	float fitness_delta_swap(CIndividual *indiv, int i, int j);
-	float fitness_delta_interchange(CIndividual *indiv, int i, int j);
-	float fitness_delta_insert(CIndividual *indiv, int i, int j);
+	double fitness_delta_swap(CIndividual *indiv, int i, int j);
+	double fitness_delta_interchange(CIndividual *indiv, int i, int j);
+	double fitness_delta_insert(CIndividual *indiv, int i, int j);
 
 
 
 private:
-    float _Evaluate(int *permu);
+    double _Evaluate(int *permu);
 };
 #endif
